Remove prowlarr type file
