--- prowlarr/type.go
@@ -1,38 +1,0 @@
-package prowlarr
-
-import (
-	"time"
-)
-
-// SystemStatus is the /api/v1/system/status endpoint.
-type SystemStatus struct {
-	Version                string    `json:"version"`
-	BuildTime              time.Time `json:"buildTime"`
-	IsDebug                bool      `json:"isDebug"`
-	IsProduction           bool      `json:"isProduction"`
-	IsAdmin                bool      `json:"isAdmin"`
-	IsUserInteractive      bool      `json:"isUserInteractive"`
-	StartupPath            string    `json:"startupPath"`
-	AppData                string    `json:"appData"`
-	OsName                 string    `json:"osName"`
-	OsVersion              string    `json:"osVersion"`
-	IsNetCore              bool      `json:"isNetCore"`
-	IsMono                 bool      `json:"isMono"`
-	IsLinux                bool      `json:"isLinux"`
-	IsOsx                  bool      `json:"isOsx"`
-	IsWindows              bool      `json:"isWindows"`
-	IsDocker               bool      `json:"isDocker"`
-	Mode                   string    `json:"mode"`
-	Branch                 string    `json:"branch"`
-	Authentication         string    `json:"authentication"`
-	DatabaseType           string    `json:"databaseType"`
-	DatabaseVersion        string    `json:"databaseVersion"`
-	MigrationVersion       int       `json:"migrationVersion"`
-	URLBase                string    `json:"urlBase"`
-	RuntimeVersion         string    `json:"runtimeVersion"`
-	RuntimeName            string    `json:"runtimeName"`
-	StartTime              time.Time `json:"startTime"`
-	PackageVersion         string    `json:"packageVersion"`
-	PackageAuthor          string    `json:"packageAuthor"`
-	PackageUpdateMechanism string    `json:"packageUpdateMechanism"`
-}
